requirements.txt
python-telegram-bot==21.7
httpx==0.27.0
Pillow==10.4.0
python-dotenv==1.0.1
